# feat: Improve ToString performance for Core.Mathematics structs