--- a/sources/core/Stride.Core.Mathematics/Color3.cs
+++ b/sources/core/Stride.Core.Mathematics/Color3.cs
@@ -27,7 +27,8 @@
 * THE SOFTWARE.
 */
 
-using System.Globalization;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace Stride.Core.Mathematics;
@@ -38,10 +39,8 @@ namespace Stride.Core.Mathematics;
 [DataContract("Color3")]
 [DataStyle(DataStyle.Compact)]
 [StructLayout(LayoutKind.Sequential, Pack = 4)]
-public struct Color3 : IEquatable<Color3>, IFormattable
+public struct Color3 : IEquatable<Color3>, ISpanFormattable
 {
-    private const string ToStringFormat = "R:{0} G:{1} B:{2}";
-
     /// <summary>
     /// The red component of the color.
     /// </summary>
@@ -710,17 +709,6 @@ public struct Color3 : IEquatable<Color3>, IFormattable
     }
 
     /// <summary>
-    /// Returns a <see cref="string"/> that represents this instance.
-    /// </summary>
-    /// <returns>
-    /// A <see cref="string"/> that represents this instance.
-    /// </returns>
-    public override readonly string ToString()
-    {
-        return ToString(CultureInfo.CurrentCulture);
-    }
-
-    /// <summary>
     /// Convert this color to an equivalent <see cref="Color4"/> with an opaque alpha.
     /// </summary>
     /// <returns>An equivalent <see cref="Color4"/> with an opaque alpha.</returns>
@@ -732,26 +720,10 @@ public struct Color3 : IEquatable<Color3>, IFormattable
     /// <summary>
     /// Returns a <see cref="string"/> that represents this instance.
     /// </summary>
-    /// <param name="format">The format.</param>
     /// <returns>
     /// A <see cref="string"/> that represents this instance.
     /// </returns>
-    public readonly string ToString(string? format)
-    {
-        return ToString(format, CultureInfo.CurrentCulture);
-    }
-
-    /// <summary>
-    /// Returns a <see cref="string"/> that represents this instance.
-    /// </summary>
-    /// <param name="formatProvider">The format provider.</param>
-    /// <returns>
-    /// A <see cref="string"/> that represents this instance.
-    /// </returns>
-    public readonly string ToString(IFormatProvider? formatProvider)
-    {
-        return string.Format(formatProvider, ToStringFormat, R, G, B);
-    }
+    public override readonly string ToString() => $"{this}";
 
     /// <summary>
     /// Returns a <see cref="string"/> that represents this instance.
@@ -761,15 +733,29 @@ public struct Color3 : IEquatable<Color3>, IFormattable
     /// <returns>
     /// A <see cref="string"/> that represents this instance.
     /// </returns>
-    public readonly string ToString(string? format, IFormatProvider? formatProvider)
+    public readonly string ToString([StringSyntax(StringSyntaxAttribute.NumericFormat)] string? format, IFormatProvider? formatProvider)
     {
-        if (format == null)
-            return ToString(formatProvider);
+        var handler = new DefaultInterpolatedStringHandler(8, 3, formatProvider);
+        handler.AppendLiteral("R:");
+        handler.AppendFormatted(R, format);
+        handler.AppendLiteral(" G:");
+        handler.AppendFormatted(G, format);
+        handler.AppendLiteral(" B:");
+        handler.AppendFormatted(B, format);
+        return handler.ToStringAndClear();
+    }
 
-        return string.Format(formatProvider, ToStringFormat,
-                             R.ToString(format, formatProvider),
-                             G.ToString(format, formatProvider),
-                             B.ToString(format, formatProvider));
+    bool ISpanFormattable.TryFormat(Span<char> destination, out int charsWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
+    {
+        var format1 = format.Length > 0 ? format.ToString() : null;
+        var handler = new MemoryExtensions.TryWriteInterpolatedStringHandler(8, 3, destination, provider, out _);
+        handler.AppendLiteral("R:");
+        handler.AppendFormatted(R, format1);
+        handler.AppendLiteral(" G:");
+        handler.AppendFormatted(G, format1);
+        handler.AppendLiteral(" B:");
+        handler.AppendFormatted(B, format1);
+        return destination.TryWrite(ref handler, out charsWritten);
     }
 
     /// <summary>