… build/Stride.sln
… sources/core/Stride.Core.Mathematics.Tests/TestToString.cs
… sources/core/Stride.Core.Mathematics/ColorHSV.cs
… sources/core/Stride.Core.Mathematics/Double2.cs
… sources/core/Stride.Core.Mathematics/Double3.cs
… sources/core/Stride.Core.Mathematics/Double4.cs
… sources/core/Stride.Core.Mathematics/Half.cs
… sources/core/Stride.Core.Mathematics/Half2.cs
… sources/core/Stride.Core.Mathematics/Half3.cs
… sources/core/Stride.Core.Mathematics/Half4.cs
… sources/core/Stride.Core.Mathematics/Int2.cs
… sources/core/Stride.Core.Mathematics/Int3.cs
… sources/core/Stride.Core.Mathematics/Int4.cs
… sources/core/Stride.Core.Mathematics/Matrix.cs
… sources/core/Stride.Core.Mathematics/Plane.cs
… sources/core/Stride.Core.Mathematics/Point.cs
… sources/core/Stride.Core.Mathematics/Quaternion.cs
… sources/core/Stride.Core.Mathematics/Ray.cs
… sources/core/Stride.Core.Mathematics/Rectangle.cs
… sources/core/Stride.Core.Mathematics/RectangleF.cs
… sources/core/Stride.Core.Mathematics/Size2.cs
… sources/core/Stride.Core.Mathematics/Size2F.cs
… sources/core/Stride.Core.Mathematics/Size3.cs
… sources/core/Stride.Core.Mathematics/UInt4.cs
… sources/core/Stride.Core.Mathematics/Vector2.cs
… sources/core/Stride.Core.Mathematics/Vector3.cs
… sources/core/Stride.Core.Mathematics/Vector4.cs