# PR Details

Implements `ISpanFormattable` for most types in `Core.Mathematics`,  inspired by this article [High-performance string formatting in .NET](https://mijailovic.net/2025/05/14/high-performance-strings/) and related [reddit post](https://www.reddit.com/r/dotnet/comments/1kmdldk/highperformance_string_formatting_in_net/).

As a side effect, a few `ToString()` implementations were fixed (which would have generated wrong values when given a custom format).

Also, there are now only two `ToString` methods:
* `ToString()` overridden from `object`
* `ToString(string?, IFormatProvider?)` implementing `IFormattable`
* `ToString(string?)` and `ToString(IFormatProvider?)` can be simulated by passing `null` for the other argument.

## Benchmarks

The gain is not significant for most types (but always positive). It is more visible for `Matrix` which has  a lot of fields.

### Matrix (before)

| Method                              | Mean       | Ratio | Allocated | Alloc Ratio |
|------------------------------------ |-----------:|------:|----------:|------------:|
| ToString                            |   859.2 ns |  1.00 |     768 B |        1.00 |
| ToString(string?)                   | 1,312.2 ns |  1.53 |     960 B |        1.25 |
| ToString(IFormatProvider?)          |   987.5 ns |  1.15 |     768 B |        1.00 |
| ToString(string?, IFormatProvider?) | 1,000.7 ns |  1.16 |     792 B |        1.03 |

### Matrix (after)

| Method                              | Mean     | Ratio | Allocated | Alloc Ratio |
|------------------------------------ |---------:|------:|----------:|------------:|
| ToString                            | 678.0 ns |  1.00 |     232 B |        1.00 |
| ToString(string?)                   | 997.9 ns |  1.47 |     376 B |        1.62 |
| ToString(IFormatProvider?)          | 666.4 ns |  0.98 |     312 B |        1.34 |
| ToString(string?, IFormatProvider?) | 997.6 ns |  1.47 |     376 B |        1.62 |

### Side by side comparison

| Method                                  | Mean       | Ratio | Allocated | Alloc Ratio |
|-----------------------------------------|-----------:|------:|----------:|------------:|
| **ToString**                            |   **859.2 ns** |  **1.00** |     **768 B** |        **1.00** |
| ToString                                |       678.0 ns |      0.79 |         232 B |            0.30 |
| **ToString(string?)**                   | **1,312.2 ns** |  **1.00** |     **960 B** |        **1.00** |
| ToString(string?)                       |       997.9 ns |      0.76 |         376 B |            0.39 |
| **ToString(IFormatProvider?)**          |   **987.5 ns** |  **1.00** |     **768 B** |        **1.00** |
| ToString(IFormatProvider?)              |       666.4 ns |      0.67 |         312 B |            0.41 |
| **ToString(string?, IFormatProvider?)** | **1,000.7 ns** |  **1.00** |     **792 B** |        **1.00** |
| ToString(string?, IFormatProvider?)     |       997.6 ns |      1.00 |         376 B |            0.47 |

*Note: in these benchmarks the format `"0.##"` and the provider `CultureInfo.CurrentCulture` are passed as parameters which explains why the benchmark for `ToString(string?, IFormatProvider?)` is slightly worse than `ToString(IFormatProvider?)` since the custom format causes more computation and allocations. If both are `null` the performance is similar to `ToString()`.*

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My change requires a change to the documentation.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] **I have built and run the editor to try this change out.**