@Kryptos-FR 2025-05-15
@Eideren do you think I should add unit tests to check the result of calling `ToString()` for these types?