@Eideren 2025-05-15
@Kryptos-FR maybe just for round-trippable values ?